Refactor fhir models to extend generic base class
